requirements.txt
bleak
pyserial-asyncio
pytest